mqtt_client/requirements.txt
annotated-types==0.5.0
click==8.1.6
dnspython==2.4.2
h11==0.14.0
mypy==1.5.0
mypy-extensions==1.0.0
paho-mqtt==1.6.1
pydantic==2.1.1
pydantic-core==2.4.0
pydantic-settings==2.0.3
pymongo==4.4.1
python-dotenv==1.0.0
tomli==2.0.1
types-paho-mqtt==1.6.0.7
typing-extensions==4.7.1
